// src/app/core/api/models/match.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MatchStatus } from '../models/match-status';
import { Team } from '../models/team';
export interface Match {
  awayGoals?: number;
  awayTeam?: Team;
  awayTeamId: number;
  homeGoals?: number;
  homeTeam?: Team;
  homeTeamId: number;
  id: number;
  kickoffTime: string;
  matchdayId: number;
  status: MatchStatus;
}